+++ source/Calamari.Tests/KubernetesFixtures/Helm3UpgradeFixture.cs
@@ -7,7 +7,6 @@ using Calamari.Testing.Helpers;
 using Calamari.Testing.Requirements;
 using FluentAssertions;
 using NUnit.Framework;
-using Octopus.Calamari.Contracts.Kubernetes;
 
 namespace Calamari.Tests.KubernetesFixtures
 {
@@ -66,14 +65,14 @@ namespace Calamari.Tests.KubernetesFixtures
             result.AssertOutputMatches("Resource Status Check: Stopped.");
 
             result.CapturedOutput.ServiceMessages
-                  .Where(sm => sm.Name == ServiceMessages.ResourceStatus.Name)
+                  .Where(sm => sm.Name == SpecialVariables.ServiceMessages.ResourceStatus.Name)
                   .Should()
                   .HaveCountGreaterOrEqualTo(1);
 
             result.CapturedOutput.ServiceMessages
-                  .Where(sm => sm.Name == ServiceMessages.ResourceStatus.Name)
+                  .Where(sm => sm.Name == SpecialVariables.ServiceMessages.ResourceStatus.Name)
                   .Should()
-                  .Contain(sm => sm.Properties[ServiceMessages.ResourceStatus.Attributes.Name] == $"mychart-configmap-{ReleaseName}" && sm.Properties[ServiceMessages.ResourceStatus.Attributes.Namespace] == Namespace && sm.Properties[ServiceMessages.ResourceStatus.Attributes.Group] == "" && sm.Properties[ServiceMessages.ResourceStatus.Attributes.Version] == "v1" && sm.Properties[ServiceMessages.ResourceStatus.Attributes.Kind] == "ConfigMap" && sm.Properties[ServiceMessages.ResourceStatus.Attributes.Status] == Kubernetes.ResourceStatus.Resources.ResourceStatus.Successful.ToString());
+                  .Contain(sm => sm.Properties[SpecialVariables.ServiceMessages.ResourceStatus.Attributes.Name] == $"mychart-configmap-{ReleaseName}" && sm.Properties[SpecialVariables.ServiceMessages.ResourceStatus.Attributes.Namespace] == Namespace && sm.Properties[SpecialVariables.ServiceMessages.ResourceStatus.Attributes.Group] == "" && sm.Properties[SpecialVariables.ServiceMessages.ResourceStatus.Attributes.Version] == "v1" && sm.Properties[SpecialVariables.ServiceMessages.ResourceStatus.Attributes.Kind] == "ConfigMap" && sm.Properties[SpecialVariables.ServiceMessages.ResourceStatus.Attributes.Status] == Kubernetes.ResourceStatus.Resources.ResourceStatus.Successful.ToString());
         }
 
         [Test]
@@ -100,7 +99,7 @@ namespace Calamari.Tests.KubernetesFixtures
                 result.AssertOutputMatches("Octopus needs Helm v3.13 or later to display object status and manifests.");
 
                 result.CapturedOutput.ServiceMessages
-                      .Where(sm => sm.Name == ServiceMessages.ResourceStatus.Name)
+                      .Where(sm => sm.Name == SpecialVariables.ServiceMessages.ResourceStatus.Name)
                       .Should()
                       .BeEmpty();
             }
@@ -129,7 +128,7 @@ namespace Calamari.Tests.KubernetesFixtures
 
             //we should not have received any KOS service messages
             result.CapturedOutput.ServiceMessages
-                  .Where(sm => sm.Name == ServiceMessages.ResourceStatus.Name)
+                  .Where(sm => sm.Name == SpecialVariables.ServiceMessages.ResourceStatus.Name)
                   .Should()
                   .BeEmpty();
         }
@@ -157,7 +156,7 @@ namespace Calamari.Tests.KubernetesFixtures
 
             //we should not have received any KOS service messages
             result.CapturedOutput.ServiceMessages
-                  .Where(sm => sm.Name == ServiceMessages.ResourceStatus.Name)
+                  .Where(sm => sm.Name == SpecialVariables.ServiceMessages.ResourceStatus.Name)
                   .Should()
                   .BeEmpty();
         }
@@ -184,7 +183,7 @@ namespace Calamari.Tests.KubernetesFixtures
 
             //we should not have received any KOS service messages
             result.CapturedOutput.ServiceMessages
-                  .Where(sm => sm.Name == ServiceMessages.ResourceStatus.Name)
+                  .Where(sm => sm.Name == SpecialVariables.ServiceMessages.ResourceStatus.Name)
                   .Should()
                   .BeEmpty();
         }
