+++ source/Calamari/ArgoCD/Conventions/ManifestTemplating/ManifestUpdateResult.cs
@@ -1,7 +1,6 @@
 #nullable enable
 using System;
 using System.Collections.Generic;
-using Octopus.Calamari.Contracts.ArgoCD;
 
 namespace Calamari.ArgoCD.Conventions.ManifestTemplating;
